Move existing class variable nodes into a package
